+++ lib/shipcloud/operations/create.rb
@@ -5,8 +5,10 @@ module Shipcloud
         # Creates a new object
         #
         # @param [Hash] attributes The attributes of the created object
-        def create(attributes)
-          response = Shipcloud.request(:post, base_url, attributes)
+        # @param \[String\] optional api_key The api key. If no api key is given, Shipcloud.api_key
+        # will be used for the request
+        def create(attributes, api_key: nil)
+          response = Shipcloud.request(:post, base_url, attributes, api_key: api_key)
           if create_response_root
             response = response.fetch(create_response_root, {})
           end