+++ lib/shipcloud/operations/find.rb
@@ -5,9 +5,11 @@ module Shipcloud
         # Finds a given object
         #
         # @param [String] id The id of the object that should be found
+        # @param \[String\] optional api_key The api key. If no api key is given, Shipcloud.api_key
+        # will be used for the request
         # @return [Shipcloud::Base] The found object
-        def find(id)
-          response = Shipcloud.request(:get, "#{base_url}/#{id}", {})
+        def find(id, api_key: nil)
+          response = Shipcloud.request(:get, "#{base_url}/#{id}", {}, api_key: api_key)
           self.new(response)
         end
       end