+++ spec/shipcloud/request/base_spec.rb
@@ -3,25 +3,26 @@ require "spec_helper"
 describe Shipcloud::Request::Base do
   context "#perform" do
     it "checks for an api key" do
-      Shipcloud.stub(:api_key).and_return(nil)
+      info = Shipcloud::Request::Info.new(:get, "shipments", nil, {})
 
       expect{
-        Shipcloud::Request::Base.new(nil).perform
+        Shipcloud::Request::Base.new(info).perform
       }.to raise_error Shipcloud::AuthenticationError
     end
 
     it "performs an https request" do
-      Shipcloud.stub(:api_key).and_return("some key")
-      connection = stub
-      validator = stub
-      Shipcloud::Request::Connection.stub(:new).and_return(connection)
-      Shipcloud::Request::Validator.stub(:new).and_return(validator)
+      info = Shipcloud::Request::Info.new(:get, "shipments", "api_key", {})
+      connection = double
+      validator = double
+      allow(Shipcloud::Request::Connection).to receive(:new).with(info).and_return(connection)
+      allow(Shipcloud::Request::Validator).to receive(:new).with(info).and_return(validator)
 
-      connection.should_receive(:setup_https)
-      connection.should_receive(:request)
-      validator.should_receive(:validated_data_for)
+      expect(connection).to receive(:setup_https)
+      response = double
+      expect(connection).to receive(:request).and_return(response)
+      expect(validator).to receive(:validated_data_for).with(response)
 
-      Shipcloud::Request::Base.new(nil).perform
+      Shipcloud::Request::Base.new(info).perform
     end
   end
 end