+++ lib/shipcloud/operations/all.rb
@@ -3,8 +3,12 @@ module Shipcloud
     module All
       module ClassMethods
         # Loads all Objects of the resource
-        def all(filter = {})
-          response = Shipcloud.request(:get, base_url, filter)
+        # @param [Hash] optional filter Filter the shipments list using one or more filter
+        # creteria
+        # @param [String] optional api_key The api key. If no api key is given, Shipcloud.api_key
+        # will be used for the request
+        def all(filter = {}, api_key: nil)
+          response = Shipcloud.request(:get, base_url, filter, api_key: api_key)
           if index_response_root
             response = response.fetch(index_response_root, [])
           end