+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## [Unreleased]
+### Added
+
+### Removed
+
+### Changed
+
+### Fixed
+
 ## [0.6.0] - 2016-01-21
 ### Added
 - This CHANGELOG file (Following "[Keep a CHANGELOG](http://keepachangelog.com/)")